requirements.txt
xmlschema
soundfile
pandas
pyyaml
yattag
bs4
audioExtractor
